Add JIT options to disable and trace constant provenance tracking
Tracking of constant provenance, i.e. of where known objects have been
found at compile time, will not be implemented in OMR, or at least there
is no plan to do so. It's up to downstream projects to decide whether
and how to do such tracking.

Unfortunately, however, in order to control the tracking based on JIT
options, it's still necessary for the options to exist in OMR.