Fix some inconsistencies with GC IDs
Before this change the IDs were being retrieved from the current count.
This lead to duplicate IDs due to the counts being retrieved and updated
 non sequentially. Another field to store the ID at the start of the
cycle is introduced. There were some inconsistencies with the increase
in count as well, due to it being updated at varying places. The count
increase is re-ordered to fix the inconsistencies.

Signed-off-by: Adrian Popescu <adpopescu@ibm.com>