Consolidate CS2 sparse bit vector copying logic
It seems that operator=(ASparseBitVector<Allocator> &) must still be
declared separately from operator=(ASparseBitVector<A2> &) to prevent
the build compiler from generating a default assignment operator. There
probably isn't a way for the more specific overload to call the less
specific overload with the same arguments, so the implementation is
moved into a new method CopyFrom().