Do not restore comp phase during comp exception
In `OMR::Optimizer::optimize()` the JIT saves the "compilation phase"
into a stack allocated object of type `TR::Compilation::CompilationPhaseScope`.
This happenes in the constructor of the object. The destructor of the
object restores the previously saved state. This ensures that that if we
leave the scope of the `OMR::Optimizer::optimize()` routine, the compilation
phase is properly restored.
However, if the optimizer throws an exception, the stack is unwound, the
destructor of the CompilationPhaseScope object is called and the compilation
phase is restored (most likely showing as ILgen). This is not want we want.
For exceptions, we want to retain the compilation phase (namely the optimization)
that produced the exception, so that it can be properly reported at higher
levels, when the exception is caught.
This commit changes the destructor of the CompilationPhaseScope object so that
the compilation phase is restored only if there isn't an exception in progress.
This ensures that the compilation phase at the time when the exception was
thrown is reported correctly.
This commit also moves the `comp()->reportOptimizationPhase(optNum);`
statement from `performOptimization()` routine to an earlier location,
as soon as we determine that the current optimization is a proper optimization
and not an optimization group. This ensures that the "analysisPhase" reported
in case of a crash or exception is shown in the context of the current optimization
and not the previous one.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>