Final copy-scan history count update
In the history table of copy-scan counts, when the final update is done,
in a simple STW Scavenge, it was not necessary to be atomic, since it
was the only GC thread to do the update.

Even with Concurrent Scavenger, where the update would not necessarily
be the final nor the only one (there could be a few attempts to converge
the scanning cycle due to non-empty copy cashes associated with mutator
threads), it would still not be a problem, since mutator threads would
not even try to update these counts, and would not race with the GC
thread doing the last few 'final' non-atomic updates.

With the introduction of taxation in Concurrent Scavenge, since mutator
threads would do scan work they would start updating the counts. Even
though their updates would be atomic, the GC thread doing non-atomic
'final' updates would create problems (like incrementing the size of the
history table beyond its real maximum).

The fix is to defer the final update from the GC thread a bit later,
when it's known the all mutator threads are done flushing.

An alternative solution is to leave the final flushes where there are,
but make them atomic. However, that seemed to be a bit more
complex/risky change (even though more future proof).

Signed-off-by: Aleksandar Micic <Aleksandar_Micic@ca.ibm.com>