Mark block cold if an original block was cold in endBlockSimplifier
Upon merging two blocks, endBlockSimplifier would mark the resultant
block as cold if the second of the original pair of blocks was cold, and
as not cold if the second was not cold.

Subsequent calls to OMR::CFG::resetFrequencies - after some later
inlining, for example - would cause a block that no longer had the
cold flag set to have its frequency updated based on profiling
information rather than remaining low frequency as the block's cold flag
had been removed.  Further, some optimizations will ignore blocks
that have the cold flag set, so its absence on a merged block would
cause such optimizations to take that block, and possibly its
successors, into account.  Either scenario could cause that block to
have an undue influence on later optimization passes.

In addition, endBlockSimplifier would only consider updating the
frequency of the resultant block if the second of the original pair had
only one predecessor.  However, it only checked that after the block had
already been removed from the control flow graph, which means that it
would never have any predecessors, and the frequency of the resultant
block would never be updated.

If either block merged by endBlockSimplifier is cold, mark the resultant
block as cold.  If the resultant block is cold, the frequency will be
the greater of the frequencies of the original pair of blocks;
otherwise, it will be the lesser of the frequencies of the original
pair.  This ensures that a path that was explicitly marked as cold
remains cold, while appropriately adjusting frequencies for blocks on
warmer paths.

Also removed a call to Block::inheritBlockInfo, which was largely
redundant.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>