Prevent cloning blocks that end with a guard in Loop Replicator
If a block that ends with a guard has been cloned by Loop Replicator,
it might interfere with later passes of Loop Versioner.  Prevent
transforming a loop whose entry block ends with a guard, and stop
cloning blocks just before the point at which a path reaches a block
that ends with a guard.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>