DDR: Emit proper base type for Type superclass
When dealing with (but perhaps not limited to) the split dwarf debug
files (.dwo) produced with Open XL on z/OS, there are instances where
the inheritance type points to a typedef Die. This updates the generator
classes (Superset and Blobgen) to look for the proper base type when
attempting to affix the supertypes.

Also changes ClassUDT::_superClass to `Type` to avoid casts.

Signed-off-by: Gaurav Chaudhari <Gaurav.Chaudhari@ibm.com>