Add nodeRegSignExtendedTo64Bit flag on node
The flag can be used by the evaluator which needs to sign extend the register.
By having this flag, we can stop using AlreadySignExtended flag on the register
flag to skip extension in the codegen. Limitation of such flag on register is
that, once the node with which the register was originally associated with is
not referenced, it can be used by other evaluator and can be associcated with
other nodes. In such cases, it can inherit the AlreadySignExtended flag which
can lead to incorrectly skipping the sign extension (For example while
evaluating i2l). Ideally we should have a mechanism so that AlreadySignExtended
flag is reset when clobber evaluating a node, but there are multiple places
where we call evaluate on the node and check if the node registers can be
clobbered or not and use the register as clobberable register if we can.
Ideally we should use the gprClobberEvaluate in all such places.

Currently the register flag is only used on IBM Z.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>