Use nodeRegSignExtendedTo64Bit flag on node for extendcast on IBM Z
On extendCastEvaluator on IBM Z, check if the flag
nodeRegSignExtendedTo64Bit is set on the node or not to skip the sign
extension. Changes in this commit sets the flag on the child node of the
regStore node when it sign extends the register so that subsequent sign
extension on the child node can be skipped.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>