Re-enable addKnownObjectConstraints() for JITServer
The functionality of addKnownObjectConstraints() was disabled for
JITServer due to the large number of messages it generates to
retrieve information related to "known objects". This problem
has been solved to a large extent in OpenJ9 through the addition
of caching (see https://github.com/eclipse-openj9/openj9/pull/23012).
This commit re-enables addKnownObjectConstraints() for JITServer.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>